gcp_devops_agent/requirements.txt
a2a-adk
fastapi
uvicorn
google-cloud-run